fix(go/adbc/driver/snowflake): try to suppress stray logs (#2608)
It turns out gosnowflake also exposes a log level control; set this to
"warn".

I also filed https://github.com/snowflakedb/gosnowflake/pull/1332. We
need the fix to be upstream since ultimately the log message is
triggered by constructing a global, so we're powerless to suppress it.
It was ultimately fixed by
https://github.com/snowflakedb/gosnowflake/pull/1327.