docs: rework "What exactly is ADBC?" in FAQ (#2763)
This PR is a bit of a check on my understanding. There are a few changes
here:

1. Changed first sentence to help draw the reader in a bit better.
2. Changes the language "a set of abstract APIs" around. Is it more
accurate to say there's one abstract API and that there are interface
definitions in a few languages, most recently including Rust?
3. Remove the "For example..." as it felt out of place and slows the
reader down from getting to bullet point 2.
4. Instead of referencing an easily out of date list of implementations,
just link to the driver status page.